Feat: WatchInterface class. It was created in the phone application a clas which is charge of managing the notification from the phone to the wearable. This class perform the following action:
Verifies if the weather condition has changed since the previous notification and stores into the shrared preferences the sent values.
This class is called by the sunshineSynctask within the syncWeather method:

/*
*Notify the wearable the new weather conditions
* In order to avoid sending notifications to the wearable anytime the SyncTask runs it is verified if the weather conditions has changed.
*/
WatchInterface watchInterface=new WatchInterface(context);
    if (watchInterface.weatherHasChanged()){
        watchInterface.notifyWearable();
	}

In case the weather condition has changed it will send to the wearable the min and Max temp already formated as an String and the weather_id as an Integer.