Feat: Show the current weather information on the wearable WatchInterfaceSevice: It was created in the wear application a clas which is charge of receiving the weather updates from the phone application. This class extends from perform the WearableListenerService and it implements the onDataChanged method. In order to provide the retrieved weather infoemtion to the watchFace it uses SharedPreferences. This class save the received wether information from the Shared Preferences and it also gets that inforation using static methods.
In order to enable the service to listen the notification it was added into the Manifest declaration and it was added the DATA_CHANGED:
<intent-filter>
                <action android:name="com.google.android.gms.wearable.DATA_CHANGED" />

                <data
                    android:host="*"
                    android:scheme="wear"
                    />
</intent-filter>

DigitalWatchFaceService:
the onDraw method was updated to check if the wearable has received weather information. In case it haven¡t received yet, the weather icon and the temperatures are not shown.
it calls the WatchInterfaceSevice's public static methods to get the received values.

#Rubric:

App displays the high and low temperatures.
App displays a graphic that summarizes the day’s weather (e.g., a sunny image, rainy image, cloudy image, etc.).
Watch face obtains weather data by talking with a copy of Sunshine running on a device paired to the watch.